Move WhatsAppForm to components/wa/
